Calculate the dynamic local trend association measure

Calculate the dynamic local trend association measure (DLTAM) for a pair of time series.

Usage

dltam(.data, x, y, timestamp = NULL, window_size, mat_seq_size)

Arguments

.data

A data frame or a data frame extension (e.g. a tibble).

x, y

Names of the columns containing the time series data, as symbols.

timestamp

An optional parameter specifying the name of the column, as a symbol, containing timestamps corresponding to x and y.

window_size

The size of the window to use when calculating the Moving Approximation Transform slope values from the time series values.

mat_seq_size

The size of the window to use when calculating Local Trend Association Measure values from the Moving Approximation Transform slope values.

Value

A tibble with columns timestamp (if supplied) and score.

Details

For a time series, the Moving Approximation Transform (MAT) uses a least-sqaures approach similar to what is used in simple linear regression, to calculate the slope coefficients for subsequences (windows) of a time series. The DLTAM then uses windows (of same or different size) of the MAT values for a pair of time series to calculate the Local Trend Association Measure (LTAM), which is simply a cosine similarity of the slope values. As such, this gives a representation of the local strength of co-movement and the direction of movement for time series pairs. As a cosine similarity measure, the value is bounded between -1 (strong negative co-movement) and +1 (strong positive co-movement).

References

García-López F. J., Batyrshin, I., Gelbukh, A. (2018) Dynamic Local Trend Associations in Analysis of Comovements of Financial Time Series. In: Fuzzy Logic in Intelligent System Design. NAFIPS 2017. Advances in Intelligent Systems and Computing, vol 648. Springer. doi:10.1007/978-3-319-67137-6_20.